Add mem_cmp() and mem_set_s ().
talloc: accept size_t as functions argument, but do not try to allocate more
than MEMMAXSIZE. This allow to catch up requests more than UINT_MAX minus the
header and alignment.

Add float_to_string() units and use it in vformat() with "%f".

This ftoa() algorithm is taken from nanoprintf:
  https://github.com/charlesnicholson/nanoprintf

Nanoprintf is dual-licensed under both the "Unlicense" and the
"Zero-Clause BSD" (0BSD) licenses. The intent of this dual-licensing
structure is to make nanoprintf as consumable as possible in as many
environments / countries / companies as possible without encumbering
users.

+++ src/y/pkg/libc/libc/mem/talloc.h
@@ -3,7 +3,7 @@ typedef struct __attribute__((packed)) {
   uint prev_chunk_size;
 } MemChunk;
 
-typedef void *(*MemAlloc) (uint);
+typedef void *(*MemAlloc) (size_t);
 
 #define MEM_HEADER_SIZE (sizeof (MemChunk))
 #define MEM_ALIGN_SIZE 8
